zenscanner/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dj-database-url==1.0.0
django==4.1.4
django-cors-headers==3.13.0
django-ninja==0.19.1
docker==6.0.1
gunicorn==20.1.0
h11==0.14.0
idna==3.4
importlib-metadata==5.1.0
kombu==5.2.4
minio==7.1.12
packaging==22.0
prompt-toolkit==3.0.36
psycopg2==2.9.5
pycryptodome==3.16.0
pydantic==1.10.2
pyjwt==2.6.0
python-magic==0.4.27
python-slugify==7.0.0
pytz==2022.6
redis==4.4.0
requests==2.28.1
sentry-sdk==1.11.1
setuptools==65.6.3
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
typing-extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
vine==5.0.0
wcwidth==0.2.5
websocket-client==1.4.2
zipp==3.11.0
